Auth, Per to user

+++ src/actions/actions.ts
@@ -1,31 +1,57 @@
 "use server";
 
-import { signIn, signOut } from "@/lib/auth";
+import { auth, signIn, signOut } from "@/lib/auth";
 import prisma from "@/lib/db";
 import { sleep } from "@/lib/utils";
-import { petFormSchema, petIdSchema } from "@/lib/validations";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 import bcrypt from "bcryptjs";
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { redirect } from "next/navigation";
 
 // --- USER ACTIONS ---
 
-export async function logIn(formData: FormData) {
+export async function logIn(formData: unknown) {
+  // Check if formData is a FormData type
+  if (!(formData instanceof FormData)) {
+    return {
+      message: "Invalid form data",
+    };
+  }
+
   await signIn("credentials", formData);
+
+  redirect("/app/dashboard");
 }
 
 export async function logOut() {
   await signOut({ redirectTo: "/" });
 }
 
-export async function signUp(formData: FormData) {
-  const hashedPassword = await bcrypt.hash(
-    formData.get("password") as string,
-    10
-  );
+export async function signUp(formData: unknown) {
+  // Check if formData is a FormData type
+  if (!(formData instanceof FormData)) {
+    return {
+      message: "Invalid form data",
+    };
+  }
 
+  // convert FormData to object
+  const formDataEntries = Object.fromEntries(formData.entries());
+
+  // validation
+  const validatedFormData = authSchema.safeParse(formDataEntries);
+  if (!validatedFormData.success) {
+    return {
+      message: "Invalid form data",
+    };
+  }
+
+  const { email, password } = validatedFormData.data;
+  const hashedPassword = await bcrypt.hash(password, 10);
   await prisma.user.create({
     data: {
-      email: formData.get("email") as string,
+      email,
       hashedPassword,
     },
   });
@@ -38,6 +64,8 @@ export async function signUp(formData: FormData) {
 export async function addPet(pet: unknown) {
   await sleep(1000);
 
+  const session = await checkAuth();
+
   const validatedPet = petFormSchema.safeParse(pet);
   if (!validatedPet.success) {
     return {
@@ -47,7 +75,14 @@ export async function addPet(pet: unknown) {
 
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session.user.id,
+          },
+        },
+      },
     });
   } catch (error) {
     return {
@@ -60,6 +95,10 @@ export async function addPet(pet: unknown) {
 export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
 
+  // authentication check
+  const session = await checkAuth();
+
+  // validation
   const validatedId = petIdSchema.safeParse(petId);
   const validatedPet = petFormSchema.safeParse(newPetData);
   if (!validatedPet.success || !validatedId.success) {
@@ -68,6 +107,22 @@ export async function editPet(petId: unknown, newPetData: unknown) {
     };
   }
 
+  // authorization check (user owns pet)
+  const pet = await getPetById(validatedId.data);
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized",
+    };
+  }
+
+  // database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -86,6 +141,10 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
+  // authentication check
+  const session = await checkAuth();
+
+  // validation
   const validatedId = petIdSchema.safeParse(petId);
   if (!validatedId.success) {
     return {
@@ -93,6 +152,22 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  // authorization check (user owns pet)
+  const pet = await getPetById(validatedId.data);
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized",
+    };
+  }
+
+  // database mutation
   try {
     await prisma.pet.delete({
       where: {
